code refactor to separate utils, first app creation function
